Moved HeartFavoriteToggle component.
