+++ src/js/common/components/Widgets/HeartFavoriteToggle/HeartFavoriteToggleIcon.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
+import DesignTokenColors from '../../Style/DesignTokenColors';
 
 const HeartFavorite = () => (
   <Icon>
